refactor: added typography component
