+++ src/components/Typography/index.ts
@@ -1,0 +1,3 @@
+import { Typography } from './Typography';
+
+export default Typography;
