+++ src/sections/Listings/LIstings.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { server, useQuery } from '../../lib/api/index';
 import { Listing } from './types';
 import { List, Button, Spin, Row, Col, Alert, Avatar, Skeleton } from 'antd';
+import Typography from '../../components';
 
 interface ListingsProps {
   title: string;
@@ -41,9 +42,11 @@ export const Listings: React.FC<ListingsProps> = ({ title }) => {
   }
   return (
     <div className='p-4 dark:bg-neutral-600'>
-      <h2 className='mb-4 text-4xl text-text-color-1 lg:tracking-widest dark:text-white'>
-        {title}
-      </h2>
+      <Typography
+        size={50}
+        label={title}
+      />
+
       <List
         className='bg-white lg:w-1/2'
         bordered
