+++ src/sections/index.ts
@@ -1,1 +1,3 @@
-export * from './Listings';
+import Listings from './Listings';
+
+export default Listings;
